added support for selectively enabling rabbit shovel via config

+++ kubernetes/update-stack.sh
@@ -205,14 +205,18 @@ EOF
     cp ${WAGGLE_CONFIG_DIR}/${NODE_MANIFEST_V2} configs/${NODE_MANIFEST_V2}
 
     # generate rabbitmq configs / secrets for kustomize
-    # NOTE We are temporarily disabling shovels while debugging the cloud issues.
-    # TODO Let's work on separating out deployment specific config like this from the main update stack scripts.
-#     cat > configs/rabbitmq/enabled_plugins <<EOF
-# [rabbitmq_prometheus,rabbitmq_management,rabbitmq_management_agent,rabbitmq_auth_mechanism_ssl,rabbitmq_shovel,rabbitmq_shovel_management,rabbitmq_mqtt].
-# EOF
-    cat > configs/rabbitmq/enabled_plugins <<EOF
+    # TODO unify how this is done for various node settings rather than it being a one off for the shovel.
+    if [ -e configs/no-shovel ]; then
+        echo "rabbitmq shovel is disabled"
+        cat > configs/rabbitmq/enabled_plugins <<EOF
 [rabbitmq_prometheus,rabbitmq_management,rabbitmq_management_agent,rabbitmq_auth_mechanism_ssl,rabbitmq_mqtt].
 EOF
+    else
+        echo "rabbitmq shovel is enabled"
+        cat > configs/rabbitmq/enabled_plugins <<EOF
+[rabbitmq_prometheus,rabbitmq_management,rabbitmq_management_agent,rabbitmq_auth_mechanism_ssl,rabbitmq_shovel,rabbitmq_shovel_management,rabbitmq_mqtt].
+EOF
+    fi
 
     cat > configs/rabbitmq/rabbitmq.conf <<EOF
 # server config
